Обширный рефакторинг под JPA и Spring (см.ниже): 1. Переработал логику под JPA и Spring покупателя 2. Рефакторинг Entity 3. Рефакторинг BuyerMenu 4. Создание BuyerRepository 5. Создание контракта BuyerService 6. Создание реализации BuyerServiceImpl 7. Рефакторинг точки входа в приложение 8. Изменение иерархии нахождения приложения
