+++ .idea/sqldialects.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="SqlDialectMappings">
-    <file url="file://$PROJECT_DIR$/src/main/java/dao/impl/BuyerDAOImpl.java" dialect="PostgreSQL" />
-    <file url="file://$PROJECT_DIR$/src/main/java/dao/impl/ShopDAOImpl.java" dialect="GenericSQL" />
+    <file url="file://$PROJECT_DIR$/src/main/java/ru/astondevs/shop/dao/impl/BuyerDAOImpl.java" dialect="PostgreSQL" />
+    <file url="file://$PROJECT_DIR$/src/main/java/ru/astondevs/shop/dao/impl/ShopDAOImpl.java" dialect="GenericSQL" />
   </component>
 </project>
