--- src/main/java/dao/OrderDAO.java
@@ -1,21 +1,0 @@
-package dao;
-
-import entity.Order;
-
-import java.util.List;
-
-public interface OrderDAO {
-
-    Order create(Order order);
-
-    Order read(long id);
-
-    boolean update(Order order);
-
-    boolean delete(long id);
-
-    List<Order> findAll();
-
-    List<Order> findOrderByBuyerId(long id);
-
-}
