--- src/main/java/entity/Buyer.java
@@ -1,20 +1,0 @@
-package entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class Buyer {
-
-    private long id;
-
-    private String name;
-
-    private long cardNumber;
-
-    private long saleValue;
-
-}
