+++ src/main/java/ru/astondevs/shop/entity/Order.java
@@ -1,0 +1,18 @@
+package ru.astondevs.shop.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+//Виктор
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class Order {
+
+    private long id;
+    private Buyer buyer;
+    private Shop shop;
+    private Product product;
+//    private List<Product> products;
+}
