+++ src/main/java/ru/astondevs/shop/service/BuyerService.java
@@ -1,0 +1,41 @@
+package ru.astondevs.shop.service;
+
+import ru.astondevs.shop.entity.Buyer;
+
+/**
+ * Интерфейс {@link BuyerService} определяет контракт для сервиса, отвечающего за управление покупателями.
+ * Он предоставляет методы для создания, поиска, обновления, удаления и получения всех покупателей.
+ */
+public interface BuyerService {
+
+    /**
+     * Создает нового покупателя на основе введенных пользователем данных.
+     *
+     * @return созданный объект {@link Buyer}.
+     */
+    Buyer create();
+
+    /**
+     * Находит покупателя по его ID и выводит информацию о нем.
+     * Если покупатель не найден, выводится соответствующее сообщение.
+     */
+    void read();
+
+    /**
+     * Обновляет данные существующего покупателя на основе введенных пользователем данных.
+     * Если покупатель не найден, выводится соответствующее сообщение.
+     */
+    void update();
+
+    /**
+     * Удаляет покупателя по его ID.
+     * Если покупатель не найден, выводится соответствующее сообщение.
+     */
+    void delete();
+
+    /**
+     * Выводит список всех покупателей.
+     * Если покупатели не найдены, выводится соответствующее сообщение.
+     */
+    void findAll();
+}
