+++ src/main/java/ru/astondevs/shop/service/ShopServiceMenu.java
@@ -1,0 +1,8 @@
+package ru.astondevs.shop.service;
+
+import ru.astondevs.shop.dao.ShopDAO;
+import ru.astondevs.shop.dao.impl.ShopDAOImpl;
+
+public class ShopServiceMenu {
+    static ShopDAO shopDAO = new ShopDAOImpl();
+}
