--- src/main/java/service/ShopServiceMenu.java
@@ -1,8 +1,0 @@
-package service;
-
-import dao.ShopDAO;
-import dao.impl.ShopDAOImpl;
-
-public class ShopServiceMenu {
-    static ShopDAO shopDAO = new ShopDAOImpl();
-}
