+++ src/main/resources/application.properties
@@ -1,7 +1,9 @@
-spring.application.name = shopaston
-spring.datasource.url=jdbc:postgresql://dpg-cuchopi3esus73a9qj60-a.oregon-postgres.render.com:5432/astondb
-spring.datasource.username=astondb_user
-spring.datasource.password=XPcZ0OKKDx3wMYsSG2DuRta1JceXAgrL
-#spring.datasource.url=jdbc:postgresql://localhost:5432/astondb
-#spring.datasource.username=postgres
-#spring.datasource.password=root
+spring.application.name = shopAston
+#spring.datasource.url=jdbc:postgresql://dpg-cuchopi3esus73a9qj60-a.oregon-postgres.render.com:5432/astondb
+#spring.datasource.username=astondb_user
+#spring.datasource.password=XPcZ0OKKDx3wMYsSG2DuRta1JceXAgrL
+spring.datasource.url=jdbc:postgresql://localhost:5432/shopAston
+spring.datasource.username=postgres
+spring.datasource.password=postgres
+
+spring.jpa.hibernate.ddl-auto= update
